The bug is fixed. THe whole login structure has been remade
