+++ src/Router.js
@@ -17,7 +17,7 @@ import EventsScreenContainer from './containers/EventsScreen'
 import EventDetailsScreen from './screens/Events/EventDetailsScreen'
 import ProfileScreenContainer from './containers/ProfileScreenContainer'
 import CameraScreen from './screens/Profile/CameraScreen'
-import StudentCardContainer from './containers/StudentCardContainer'
+import CompanyStudentCardContainer from './containers/CompanyStudentCardContainer'
 import StudentListContainer from './containers/StudentListContainer'
 import AboutScreenContainer from './containers/AboutScreen'
 import ArkadTeamScreenContainer from './containers/ArkadTeamScreen'
@@ -162,13 +162,13 @@ const MainStack = createBottomTabNavigator(
             screen: ProfileScreenContainer,
             navigationOptions: ({ navigation }) => ({
               ...navigationOptions,
-              title: 'Login',
+              title: 'Profile',
               header: navigation.state.params ? navigation.state.params.header : null,
               headerRight: navigation.state.params ? navigation.state.params.headerRight : undefined
             })
           },
           DetailStudent: {
-            screen: StudentCardContainer,
+            screen: CompanyStudentCardContainer,
             navigationOptions: ({ navigation }) => ({
               ...navigationOptions,
               title: navigation.state.params.item.name,
@@ -181,24 +181,6 @@ const MainStack = createBottomTabNavigator(
               ...navigationOptions,
               title: navigation.state.params.item.name,
               headerRight: navigation.state.params ? navigation.state.params.headerRight : undefined,
-            })
-          },
-          StudentLogin: {
-            screen: StudentCardContainer,
-            navigationOptions: ({ navigation }) => ({
-              ...navigationOptions,
-              title: navigation.state.params.name,
-              headerRight: navigation.state.params ? navigation.state.params.headerRight : undefined,
-              headerLeft: navigation.state.params ? navigation.state.params.headerLeft : undefined
-            })
-          },
-          CompanyLogin: {
-            screen: StudentListContainer,
-            navigationOptions: ({ navigation }) => ({
-              ...navigationOptions,
-              title: 'Profile',
-              headerRight: navigation.state.params ? navigation.state.params.headerRight : undefined,
-              headerLeft: null
             })
           },
           CameraScreen: {
